+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.14.0-rc.69",
-  genVersion: "2.731.4",
-  userAgent: "speakeasy-sdk/typescript 3.14.0-rc.69 2.731.4 2.0 @orq-ai/node",
+  sdkVersion: "3.14.0-rc.70",
+  genVersion: "2.731.6",
+  userAgent: "speakeasy-sdk/typescript 3.14.0-rc.70 2.731.6 2.0 @orq-ai/node",
 } as const;