+++ packages/orq-rc/src/models/operations/createbudget.ts
@@ -664,7 +664,7 @@ export const CreateBudgetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-10-29T09:19:25.922Z",
+    "2025-10-30T04:37:46.623Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -702,7 +702,7 @@ export const CreateBudgetResponseBody$outboundSchema: z.ZodType<
   isActive: z.boolean(),
   consumption: z.lazy(() => Consumption$outboundSchema).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-10-29T09:19:25.922Z"))
+  updated: z.date().default(() => new Date("2025-10-30T04:37:46.623Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {