+++ packages/orq-rc/RELEASES.md
@@ -4689,3 +4689,13 @@ Based on:
 - [typescript v3.14.0-rc.69] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.69] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.69 - packages/orq-rc
+
+## 2025-10-30 04:39:32
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.642.2 (2.731.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.70] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.70] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.70 - packages/orq-rc